requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.2
contourpy==1.2.0
cycler==0.12.1
deprecation==2.1.0
fonttools==4.49.0
frozenlist==1.4.1
grpcio==1.60.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
multidict==6.0.5
my-tinkoff-investments @ git+https://github.com/555Russich/my_tinkoff_investments
numpy==1.26.4
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-moex @ git+https://github.com/555Russich/MOEX_API
six==1.16.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b66
tzdata==2024.1
yarl==1.9.4
backtrader @ git+https://github.com/555Russich/backtrader@my_dev
